Add files via upload
The file *REMLHFH.BoxCox.R"* contains the REML Fisher-Scoring algorithm of the Heteroscedastic Fay-Herriot (HFH) model with the Box-Cox family transformations for the variance of the random effects.

The file *use_example.R* contains an example of fitting the HFH model to a synthetic data set.

The file *data.csv* contains the synthetic data used in the example. The columns of the data set refer to: province (dom), sex, the target variable (yd), the explanatory variable for the fixed-effects (X), the explanatory variable for the heteroscedasticity parameters (W) and the variances estimates of the direct estimators (sigmaed2).